+++ src/Surfnet/StepupMiddleware/ApiBundle/Identity/Projector/AuditLogProjector.php
@@ -23,14 +23,23 @@ use Broadway\EventHandling\EventListener;
 use DateTime as CoreDateTime;
 use Rhumsaa\Uuid\Uuid;
 use Surfnet\Stepup\DateTime\DateTime;
+use Surfnet\Stepup\Identity\AuditLog\Metadata;
 use Surfnet\Stepup\Identity\Event\AuditableEvent;
 use Surfnet\Stepup\Identity\Event\IdentityForgottenEvent;
+use Surfnet\Stepup\Identity\Event\SecondFactorVettedEvent;
+use Surfnet\Stepup\Identity\Event\SecondFactorVettedWithoutTokenProofOfPossession;
 use Surfnet\Stepup\Identity\Value\CommonName;
+use Surfnet\Stepup\Identity\Value\VettingType;
 use Surfnet\StepupMiddleware\ApiBundle\Exception\RuntimeException;
 use Surfnet\StepupMiddleware\ApiBundle\Identity\Entity\AuditLogEntry;
 use Surfnet\StepupMiddleware\ApiBundle\Identity\Repository\AuditLogRepository;
 use Surfnet\StepupMiddleware\ApiBundle\Identity\Repository\IdentityRepository;
+use function is_null;
+use function property_exists;
 
+/**
+ * @SuppressWarnings(PHPMD.CouplingBetweenObjects)
+ */
 class AuditLogProjector implements EventListener
 {
     /**
@@ -69,6 +78,8 @@ class AuditLogProjector implements EventListener
     /**
      * @param AuditableEvent $event
      * @param DomainMessage  $domainMessage
+     * @SuppressWarnings(PHPMD.CyclomaticComplexity)
+     * @SuppressWarnings(PHPMD.NPathComplexity)
      */
     private function applyAuditableEvent(AuditableEvent $event, DomainMessage $domainMessage)
     {
@@ -91,6 +102,8 @@ class AuditLogProjector implements EventListener
             $entry->actorId         = $metadata['actorId'];
             $entry->actorCommonName = $actor->commonName;
         }
+
+        $this->augmentActorCommonName($entry, $auditLogMetadata);
 
         if (isset($metadata['actorInstitution'])) {
             $entry->actorInstitution = $metadata['actorInstitution'];
@@ -129,5 +142,12 @@ class AuditLogProjector implements EventListener
 
         $this->auditLogRepository->saveAll($entriesWhereActor);
         $this->auditLogRepository->removeByIdentityId($event->identityId);
+    }
+
+    private function augmentActorCommonName(AuditLogEntry $entry, Metadata $auditLogMetadata): void
+    {
+        if (property_exists($auditLogMetadata, 'vettingType') && !is_null($auditLogMetadata->vettingType)) {
+            $entry->actorCommonName .= $auditLogMetadata->vettingType->auditLog();
+        }
     }
 }