+++ src/Surfnet/StepupMiddleware/CommandHandlingBundle/Identity/Service/SecondFactorRevocationMailService.php
@@ -26,10 +26,11 @@ use Surfnet\Stepup\Identity\Value\SecondFactorIdentifier;
 use Surfnet\StepupBundle\Value\SecondFactorType;
 use Surfnet\StepupMiddleware\CommandHandlingBundle\Configuration\Service\EmailTemplateService;
 use Surfnet\StepupMiddleware\CommandHandlingBundle\Value\Sender;
+use Surfnet\StepupMiddleware\MiddlewareBundle\Service\SecondFactorDisplayNameResolverService;
 use Swift_Mailer as Mailer;
 use Swift_Message as Message;
-use Symfony\Component\Templating\EngineInterface;
 use Symfony\Component\Translation\TranslatorInterface;
+use Twig\Environment;
 
 /**
  * @SuppressWarnings(PHPMD.CouplingBetweenObjects)
@@ -52,9 +53,9 @@ final class SecondFactorRevocationMailService
     private $translator;
 
     /**
-     * @var EngineInterface
+     * @var Environment
      */
-    private $templateEngine;
+    private $twig;
 
     /**
      * @var \Surfnet\StepupMiddleware\CommandHandlingBundle\Configuration\Service\EmailTemplateService
@@ -72,13 +73,19 @@ final class SecondFactorRevocationMailService
     private $selfServiceUrl;
 
     /**
+     * @var \Surfnet\StepupMiddleware\MiddlewareBundle\Service\SecondFactorDisplayNameResolverService
+     */
+    private $displayNameResolver;
+
+    /**
      * @param Mailer $mailer
      * @param Sender $sender
      * @param TranslatorInterface $translator
-     * @param EngineInterface $templateEngine
+     * @param Environment $twig
      * @param EmailTemplateService $emailTemplateService
      * @param string $fallbackLocale
      * @param string $selfServiceUrl
+     * @param SecondFactorDisplayNameResolverService $displayNameResolver
      *
      * @throws \Assert\AssertionFailedException
      */
@@ -86,10 +93,11 @@ final class SecondFactorRevocationMailService
         Mailer $mailer,
         Sender $sender,
         TranslatorInterface $translator,
-        EngineInterface $templateEngine,
+        Environment $twig,
         EmailTemplateService $emailTemplateService,
         $fallbackLocale,
-        $selfServiceUrl
+        $selfServiceUrl,
+        SecondFactorDisplayNameResolverService $displayNameResolver
     ) {
         Assertion::string($fallbackLocale, 'Fallback locale "%s" expected to be string, type %s given');
         Assertion::string($selfServiceUrl, 'Self Service URL "%s" expected to be string, type %s given');
@@ -97,10 +105,11 @@ final class SecondFactorRevocationMailService
         $this->mailer = $mailer;
         $this->sender = $sender;
         $this->translator = $translator;
-        $this->templateEngine = $templateEngine;
+        $this->twig = $twig;
         $this->emailTemplateService = $emailTemplateService;
         $this->fallbackLocale = $fallbackLocale;
         $this->selfServiceUrl = $selfServiceUrl;
+        $this->displayNameResolver = $displayNameResolver;
     }
 
     /**
@@ -120,7 +129,7 @@ final class SecondFactorRevocationMailService
         $subject = $this->translator->trans(
             'mw.mail.second_factor_revoked.subject',
             [
-                '%tokenType%' => $secondFactorType
+                '%tokenType%' => $this->displayNameResolver->resolveByType($secondFactorType)
             ],
             'messages',
             $locale->getLocale()
@@ -135,7 +144,7 @@ final class SecondFactorRevocationMailService
             'isRevokedByRa'   => true,
             'templateString'  => $emailTemplate->htmlContent,
             'commonName'      => $commonName->getCommonName(),
-            'tokenType'       => $secondFactorType->getSecondFactorType(),
+            'tokenType'       => $this->displayNameResolver->resolveByType($secondFactorType),
             'tokenIdentifier' => $secondFactorIdentifier->getValue(),
             'selfServiceUrl'  => $this->selfServiceUrl,
             'locale'          => $locale->getLocale(),
@@ -143,8 +152,8 @@ final class SecondFactorRevocationMailService
 
         // Rendering file template instead of string
         // (https://github.com/symfony/symfony/issues/10865#issuecomment-42438248)
-        $body = $this->templateEngine->render(
-            'SurfnetStepupMiddlewareCommandHandlingBundle:SecondFactorMailService:email.html.twig',
+        $body = $this->twig->render(
+            '@SurfnetStepupMiddlewareCommandHandling/SecondFactorMailService/email.html.twig',
             $parameters
         );
 
@@ -176,7 +185,7 @@ final class SecondFactorRevocationMailService
         $subject = $this->translator->trans(
             'mw.mail.second_factor_revoked.subject',
             [
-                '%tokenType%' => $secondFactorType
+                '%tokenType%' => $this->displayNameResolver->resolveByType($secondFactorType)
             ],
             'messages',
             $locale->getLocale()
@@ -191,7 +200,7 @@ final class SecondFactorRevocationMailService
             'isRevokedByRa'   => false,
             'templateString'  => $emailTemplate->htmlContent,
             'commonName'      => $commonName->getCommonName(),
-            'tokenType'       => $secondFactorType->getSecondFactorType(),
+            'tokenType'       => $this->displayNameResolver->resolveByType($secondFactorType),
             'tokenIdentifier' => $secondFactorIdentifier->getValue(),
             'selfServiceUrl'  => $this->selfServiceUrl,
             'locale'          => $locale->getLocale(),
@@ -199,8 +208,8 @@ final class SecondFactorRevocationMailService
 
         // Rendering file template instead of string
         // (https://github.com/symfony/symfony/issues/10865#issuecomment-42438248)
-        $body = $this->templateEngine->render(
-            'SurfnetStepupMiddlewareCommandHandlingBundle:SecondFactorMailService:email.html.twig',
+        $body = $this->twig->render(
+            '@SurfnetStepupMiddlewareCommandHandling/SecondFactorMailService/email.html.twig',
             $parameters
         );