+++ src/Surfnet/StepupMiddleware/ManagementBundle/Tests/Validator/Fixtures/invalid_reconfigure_institution_request/greater_thant_zero_number_of_tokens_per_identity_option.php
@@ -24,6 +24,7 @@ return [
             'use_ra_locations' => true,
             'show_raa_contact_information' => true,
             'verify_email' => false,
+            'self_vet' => false,
             'number_of_tokens_per_identity' => -1,
             'allowed_second_factors' => [],
         ]