+++ src/Surfnet/StepupMiddleware/ManagementBundle/Tests/Validator/Fixtures/invalid_reconfigure_institution_request/missing_use_ra_locations_option.php
@@ -23,6 +23,7 @@ return [
         'surfnet.nl' => [
             'show_raa_contact_information' => true,
             'verify_email' => false,
+            'self_vet' => false,
             'number_of_tokens_per_identity' => 1,
             'allowed_second_factors' => [],
         ]